+++ src/main/java/edu/kis/vh/nursery/collection/StackInterface.java
@@ -1,4 +1,4 @@
-package edu.kis.vh.nursery.list;
+package edu.kis.vh.nursery.collection;
 
 public interface StackInterface {
     void push(int i);
